CEPH: fix attr cache not work when disable ACL
In the CephFS FSAL export implementation, the callback function
fs_supported_attrs is not properly implemented. By default, the ATTR_ACL
flag is always set in the returned attributes. However, when the
DISABLE_ACL option is enabled in the export configuration, the ATTR_ACL
flag should be explicitly unset in the request_mask. This
inconsistency between the attribute handling in request_mask and attrs
fs_supported_attrs returned causes the mdcache_is_attrs_valid function
to always return false, preventing any cache hits for inode attributes
in the metadata cache.

Change-Id: I4f9dde91d55fd85f6ca17fd22e896194a09cba81
Signed-off-by: Shaohui Wang <wangshaohui.0512@bytedance.com>