Modify Enable_UDP config option to take NLM option
UDP needs to be enabled for NFS Ganesha to receive NLM messages
from rpc-statd.  When an NFS client reboots it sends SM_NOTIFY
to rpc-statd on the NSF Ganesha server.  Then rpc-statd needs
to be able to send an SM_UNMON to NFS Ganesha's lock manager.

If that communication is not possible, the client will keep
sending the SM_NOTIFY until it gives up.  The NFS server
instances will become unstable.

rpc-statd will only send this via UDP.  Enabling all of the
UDP port options with Enable_UDP = TRUE is not desirable.
Only the MOUNT option really needs to be specified.

NFS should not be enabled with UDP ports.  So we add the ability
to specify

Enable_UDP = MOUNT, NLM

Change-Id: Id435c32f915e6e6aef123c58edb2077ed05241f1
Signed-off-by: Peter Schwenke <pschwenke@ddn.com>