Add cid_confirmed_saved to nfs_client_id_t
This is in preparation for the Virtual IP Based recovery.

This required to be able to cleanup NFSv4.0 CLIDs on umount
for the IP based in-situ reclaim. The indicator for being able to
cleanup the CLID in fs_rm_clid() is that the client ID is in
the confirmed state.  We only want the CLID to be removed when
the file system is no longer used by any of the nodes i.e.
unmounted.

Change-Id: I7ab4a74cacaf4fb2b5054352f69c7d5de3a1daee
Signed-off-by: Peter Schwenke <pschwenke@ddn.com>