READDIR: Add new enable/disable of dir chunking with FSAL option to control
Instead of setting Dir_Chunk = 0 to disable chunking, add a new
Dir_Chunk_Enable that can be set to false.

If Dir_Chunk is set to 0, Dir_Chunk_Enable will be set to false and
Dir_Chunk restored to the default (128) for compatibility with old
config files.

Also add a configuration for readdir mode for FSALs. They can set to
ANY, CHUNK, or UNCHUNK. Either of the second forms forces the mode
for those FSALs that depend on a specific mode.

FSAL_RGW sets to CHUNK

FSAL_CEPH sets to UNCHUNK

Change-Id: I499f766dfebf8986029cdf72678223e47884935e
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>