Make FSAL readlink use utf8string instead of gsh_buffdesc and eliminate gsh_strldup
gsh_strldup was used in several places to make copies of a string in
either a gsh_buffdesc or a utf8string which use different types for the
length. Also the length in gsh_buffdesc is the length of the buffer so
it is string length +1. The length was not being consistently set.

Changing the FSAL readlink method to take a utf8string instead of a
gsh_buffdesc allows for a clear definitions since we assure that
internally all utf8strings are NUL terminated with the length just
being a conveniemce.

Introduced a new function that copies a character buffer (that need not
be NUL terminated) into a utf8string, making a copy of the buffer,
adding a NUL termination, and setting the utf8string_len properly.

This allows elimination of all uses of the gsh_strldup macro that
doesn't convert to an inline function well (due the the unknown type
of the length variable) as part of a subsequent patch to make more
header files usable with C++.

Change-Id: Ifd5a250ba500f005d4345cac0fd1defaa14f8485
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>